more tests working

+++ tensorflow/lite/kernels/internal/reference/integer_ops/add.h
@@ -37,8 +37,8 @@ inline void AddElementwise(int size, const ArithmeticParams& params,
   for (int i = 0; i < size; ++i) {
     const int32 input1_val = params.input1_offset + input1_data[i];
     const int32 input2_val = params.input2_offset + input2_data[i];
-    const int32 shifted_input1_val = input1_val * (1 << params.left_shift);
-    const int32 shifted_input2_val = input2_val * (1 << params.left_shift);
+    const int32 shifted_input1_val = input1_val * (static_cast<int32_t>(1) << params.left_shift);
+    const int32 shifted_input2_val = input2_val * (static_cast<int32_t>(1) << params.left_shift);
     const int32 scaled_input1_val =
         MultiplyByQuantizedMultiplierSmallerThanOneExp(
             shifted_input1_val, params.input1_multiplier, params.input1_shift);
@@ -110,9 +110,9 @@ inline void BroadcastAdd4DSlow(const ArithmeticParams& params,
               params.input2_offset +
               input2_data[SubscriptToIndex(desc2, b, y, x, c)];
           const int32_t shifted_input1_val =
-              input1_val * (1 << params.left_shift);
+              input1_val * (static_cast<int32_t>(1) << params.left_shift);
           const int32_t shifted_input2_val =
-              input2_val * (1 << params.left_shift);
+              input2_val * (static_cast<int32_t>(1) << params.left_shift);
           const int32_t scaled_input1_val =
               MultiplyByQuantizedMultiplierSmallerThanOneExp(
                   shifted_input1_val, params.input1_multiplier,
